requirements.txt
vosk
soundcard
pillow
requests
bs4
opencv-python
deep_translator